cast both operands to string
`getAttribute()` returns a `string` or `null`, so we cast the left
operand to a string to handle `null`.

since we are using a strict comparison, we also cast the right operand
to a string.

the most common issue this solves is when the user tries to `select` an
integer.

fixes #134