Fix the page stub
Hi,
I submitted this pull request a couple weeks ago referencing Issue #496, but it seems that it was automatically rejected because I forgot to add a description. So here it is again.

The stub as it is now extends `Laravel\Dusk\Page` instead of the abstract class `Tests\Browser\Pages\Page` included with a fresh installation of Dusk.  This means that any custom shortcuts defined in `Tests\Browser\Pages\Page::siteElements()` aren't available to tests.

This has already broken one of my tests when I forgot to change the `extends` for a page.  So I'm submitting a PR changing the extends in this stub. Thanks for looking into this.