Add ChromeDriver command
This command will allow users to update their chrome driver for their OS. It has a flag that will also allow users to update every binary if they want. I've added this to the install command so the driver is updated whenever the install command is run. This means we don't have to update the docs.

I wrote this as an install command so we can remove the binaries in a future version of Dusk and always have them installed with this command. This will save having them to be packaged with the repo. It's also written in such a way that it only installs the binary for the current OS.

Fixes https://github.com/laravel/dusk/issues/641