fit content to `<html>` instead of `<body>`
most browsers apply a margin of 8px to the `<body>` tag automatically. it's very common for most CSS frameworks to "normalize" this to zero, which is why for most people the <body> size and the <html> size are the same.

since it's not the case for everyone, we can use the <html> size here instead.