assertVue*() methods support Vue 3 composition API
If Vue's composition API used, reactive data and computed props cannot
be inspected through these methods.

When fetching a component prop, attempt these paths:

1. el.__vue__.key
   * Vue 2
2. el.__vueParentComponent.ctx.key
   * Vue 3 options API
   * Vue 3 composition API for definedProps()
3. el.__vueParentComponent.setupState.key
   * Vue 3 composition API reactive(), computed(), or any other
     <script setup> variables.

If the ctx key is undefined, the setupState fallback is attempted.
Wrap in a try/catch for attempts to access undefined nested props which
could throw an error.